update: update path after teacher login
update path after teacher login
