+++ frontend/src/App.tsx
@@ -10,7 +10,6 @@ import {
 import '@/assets/styles/global.css';
 
 const HomepageStudent = lazy(() => import('@/pages/student/Homepage'));
-const HomePageTeacher = lazy(() => import('@/pages/teacher/Homepage'));
 const ClassesPage = lazy(() => import('@/pages/teacher/Classes'));
 const AdminDashboard = lazy(() => import('@/pages/admin/Dashboard'));
 const LoginTeacher = lazy(() => import('@/pages/teacher/Login'));
@@ -28,7 +27,6 @@ const App: React.FC = () => {
 				<Routes>
 					<Route index element={<Navigate to="/student" replace />} />
 					<Route path="/student" element={<HomepageStudent />} />
-					<Route path="/teacher" element={<HomePageTeacher />} />
 					<Route path="/teacher/login" element={<LoginTeacher />} />
 					<Route path="/teacher/classes" element={<ClassesPage />} />
 					<Route path="/teacher/classes" element={<ClassesPage />} />
